Start of network rewrite
Started rewriting the way networking works and moved all files related to networking to a new folder.

Started working on a join handshake
When a client joins it now gets hint data from server
When any player finds a report it displays the report hint for all players

Still lots to do, like adding a time limit on the handshake and syncing item data when a player joins. I may also want to reorganize the network files again as it is kind of messy and the way I have made most of it extend off the main window class doesn't sit well with me.